Coerce default values
Ensure that default values are coerced appropriately when used.

If the value (or default) cannot be coerced, then a new error,
CoercianError, is raised. This allows us to capture any / all
errors that may rise - e.g. ValueError, TypeError, InvalidOperation,
and so on.

I have also bumped the version to 0.3.0 - it's a minor bump rather
than a patch, as although the public interface is unchanged, the
change in errors raised may cause some try..except to fail.